feat: port api .js to next
